+++ Domain/CashDesk/Sale.cs
@@ -3,7 +3,7 @@
 public class Sale
 {
     public List<SaleItem> Items { get; private set; }
-    public int Total { get; private set; }
+    public long Total { get; private set; }
     
     public Sale()
     {
